+++ nengo/networks/circularconvolution.py
@@ -88,7 +88,7 @@ def dft_half(n):
 
 
 def CircularConvolution(n_neurons, dimensions, invert_a=False, invert_b=False,
-                        input_magnitude=1, net=None):
+                        input_magnitude=1.0, net=None):
     """Compute the circular convolution of two vectors.
 
     The circular convolution `c` of vectors `a` and `b` is given by
@@ -107,7 +107,7 @@ def CircularConvolution(n_neurons, dimensions, invert_a=False, invert_b=False,
     Parameters
     ----------
     n_neurons : int
-        Number of neurons to be used, in total.
+        Number of neurons to use in each product computation
     dimensions : int
         The number of dimensions of the input and output vectors.
     invert_a, invert_b : bool